18.5.Saving a ManyToMany Relation + Joins{Rendering the Tags}
Rendering the Tags
And now we can turn back to building our site: open the article/show.html.twig template. On this page, let's print the tags right under the article title. So, scroll down a bit. Copy the span for the heart count and paste it below.
Because the Article object holds an array of tags, use for tag in article.tags
Inside, let's create a cute little badge and print the tag name: {{ tag.name }}.
Super cool! Try it: find the page and refresh. We got it! Well, this Article only has one tag - boring. Find a different one: boom! Four different tags.
Repeat this on the homepage: we'll list the tags right under the title. Copy the for loop, then open homepage.html.twig. Down below, add a <br>, then paste! Wrap this in a <small> tag and change the class to badge-light.
This is just the same thing again: we have an article variable, which allows us to easily loop over its tags.
But notice, before we refresh, there are 8 queries. But now... there are 15! The page works, but we have another N+1 query problem. And, it's probably no big deal, but let's learn how to add a JOIN to a ManyToMany query so that we can fix it.